AUTO

+++ src/minecraft/pixlepix/missioncontrol/common/helper/LaserFX.java
@@ -32,7 +32,7 @@ public class LaserFX extends EntityFX {
     
     
     public float calculateDistance(float x1,float y1,float z1,float x2,float y2, float z2){
-    	return Math.sqrt(Math.pow(x1-x2, 2)+(Math.pow(y1-y2, 2)+(Math.pow(z1-z2, 2));
+    	return (float)Math.sqrt(Math.pow(x1-x2, 2))+(Math.pow(y1-y2, 2))+(Math.pow(z1-z2, 2));
     	
     }
